Make generate, add file exist condition

+++ parts/linux/cloud-init/artifacts/update_certs.sh
@@ -5,12 +5,7 @@ set -x
 certSource=/opt/certs
 certDestination=/usr/local/share/ca-certificates/certs
 
-if [[ -d "$certSource" && -d "$certDestination" ]]; then
-  cp -a "$certSource"/. "$certDestination"
-else
-  echo "Custom CA Trust directories not present (feature most likely not enabled), exiting"
-  exit 0
-fi
+cp -a "$certSource"/. "$certDestination"
 
 if [[ -z $(ls -A "$certSource") ]]; then
   ls "$certDestination" | grep -E '^[0-9]{14}' | while read -r line; do
@@ -22,7 +17,7 @@ else
   currIterationTag=${currIterationCertFile:0:14}
   for file in "$certDestination"/*.crt; do
       currFile=${file##*/}
-     if [[ "${currFile:0:14}" != "${currIterationTag}" ]]; then
+     if [[ "${currFile:0:14}" != "${currIterationTag}" && -f "${file}" ]]; then
           rm "${file}"
      fi
   done
